# openapi.yaml
openapi: 3.1.0
...
info:
  title: Clinical Evidence Safe Search
  description: >
    Evidence gateway for dentistry. Searches only approved scientific,
    regulatory, professional, and manufacturer sources.
    All source allowlisting and validation is enforced server-side, so the
    model cannot widen the source set: a domain absent from the allowlist
    can be neither fetched nor returned.

    Key guarantees. (1) Manufacturer documents always come back as
    MANUFACTURER_INFORMATION and can never be used to establish clinical
    superiority. (2) Laboratory and bench studies always come back as
    EARLY_PRECLINICAL with clinical_translation=uncertain. (3) Regulatory
    status is never presented as clinical effectiveness, and FDA clearance
    is never conflated with FDA approval. (4) Absent metadata is null,
    never guessed. (5) Retracted work is flagged and excluded from the
    recommendable set.
  version: 1.0.0
servers:
  - url: https://evidence.example.com
    description: Replace with your deployed HTTPS base URL
paths:
...
  /v1/source/verify:
    post:
      operationId: verifyClinicalSource
      summary: Verify whether a source is approved
      description: >
        Check whether a URL is on the server-side hard allowlist, what
        category of source it is, and what it may and may not be used
        for. Use this whenever a user pastes a link or names a website.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: Absolute URL to check.
              required:
                - url
      responses:
        '200':
          description: Source validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceVerifyResponse'
...
components:
  securitySchemes:
    ClinicalAPIKey:
      type: apiKey
      in: header
      name: X-Clinical-Key
  schemas:
...
    SourceVerifyResponse:
      type: object
      properties:
        allowed:
          type: boolean
        domain:
          type: [string, "null"]
        source_category:
          type: string
        trust_tier:
          type: [string, "null"]
        allowed_use:
          type: string
        forbidden_use:
          type: [string, "null"]
        reason:
          type: string
...
security:
  - ClinicalAPIKey: []